refactor(driver): split scmi_server.c into per-protocol files
