+++ driver/virtio/scmi/server.c
@@ -1,0 +1,6 @@
+/* Clock, reset, and power domain protocol handlers each live in
+ * their own compilation unit:
+ *   clock.c  - hvisor_scmi_clock_ioctl()
+ *   reset.c  - hvisor_scmi_reset_ioctl()
+ *   power.c  - hvisor_scmi_power_ioctl()
+ */
